Add new API module, logger module, model module, request module, and vectorizer module to the project.
This commit adds new files to the project: `api.py`, `logger.py`,
`model.py`, `request_module.py`, and `vectorizer.py`.

The `api.py` file contains the definition of the `Api` class, which is
used to create and configure the FastAPI app. It also includes two
endpoint functions: `read_root()` and `predictions()`.

The `logger.py` file contains a class `CustomFormatter` that formats the
log output using custom colors and styles. It also configures the
logging level and output destination based on settings in the
`pyproject.toml` file.

The `model.py` file contains the definition of the `Model` class, which
loads and uses the trained model for making predictions. It includes
methods for loading models from disk, making predictions, and writing
the predicted class to an output file.

The `request_module.py` file contains a function `download_model_en()`
that downloads the English model file from a remote location and stores
it locally in the `models` directory.

The `vectorizer.py` file contains a function `sentence_to_vec()` that
transforms a sentence into a vector representation using pre-trained
word embeddings. It also includes a class `GloVeVectorizer` that
converts a list of sentences into their respective vector
representations using the GloVe word embeddings.

The `main.py` file is added to the project, but it is empty at the
moment.

These changes enable the project to have an API module, logger module,
model module, request module, and vectorizer module, which will be used
for various functionalities such as handling API requests, logging,
making predictions, downloading models, and vectorizing sentences.